refactor(schemavalidate): split number-instance tests; wrap unfit-literal cause (#282)
